+++ src/pocketmine/block/Banner.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace pocketmine\block;
 
 use Ds\Deque;
+use pocketmine\block\tile\Banner as TileBanner;
 use pocketmine\block\utils\BannerPattern;
 use pocketmine\block\utils\BlockDataValidator;
 use pocketmine\block\utils\DyeColor;
@@ -34,7 +35,6 @@ use pocketmine\math\AxisAlignedBB;
 use pocketmine\math\Facing;
 use pocketmine\math\Vector3;
 use pocketmine\Player;
-use pocketmine\tile\Banner as TileBanner;
 use function assert;
 use function floor;
