+++ src/pocketmine/block/Skull.php
@@ -23,6 +23,7 @@ declare(strict_types=1);
 
 namespace pocketmine\block;
 
+use pocketmine\block\tile\Skull as TileSkull;
 use pocketmine\block\utils\BlockDataValidator;
 use pocketmine\block\utils\SkullType;
 use pocketmine\item\Item;
@@ -32,7 +33,6 @@ use pocketmine\math\AxisAlignedBB;
 use pocketmine\math\Facing;
 use pocketmine\math\Vector3;
 use pocketmine\Player;
-use pocketmine\tile\Skull as TileSkull;
 use function assert;
 use function floor;
