+++ src/pocketmine/block/tile/Banner.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\tile;
+namespace pocketmine\block\tile;
 
 use Ds\Deque;
 use pocketmine\block\utils\BannerPattern;
