+++ src/pocketmine/block/tile/Comparator.php
@@ -21,21 +21,40 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\tile;
+namespace pocketmine\block\tile;
 
+use pocketmine\block\RedstoneComparator;
 use pocketmine\nbt\tag\CompoundTag;
 
-class EnderChest extends Spawnable{
+/**
+ * @deprecated
+ * @see RedstoneComparator
+ */
+class Comparator extends Tile{
+	private const TAG_OUTPUT_SIGNAL = "OutputSignal"; //int
+
+	/** @var int */
+	protected $signalStrength = 0;
+
+	/**
+	 * @return int
+	 */
+	public function getSignalStrength() : int{
+		return $this->signalStrength;
+	}
+
+	/**
+	 * @param int $signalStrength
+	 */
+	public function setSignalStrength(int $signalStrength) : void{
+		$this->signalStrength = $signalStrength;
+	}
 
 	public function readSaveData(CompoundTag $nbt) : void{
-
+		$this->signalStrength = $nbt->getInt(self::TAG_OUTPUT_SIGNAL, 0, true);
 	}
 
 	protected function writeSaveData(CompoundTag $nbt) : void{
-
-	}
-
-	protected function addAdditionalSpawnData(CompoundTag $nbt) : void{
-
+		$nbt->setInt(self::TAG_OUTPUT_SIGNAL, $this->signalStrength);
 	}
 }
