+++ src/pocketmine/block/tile/NameableTrait.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\tile;
+namespace pocketmine\block\tile;
 
 use pocketmine\item\Item;
 use pocketmine\nbt\tag\CompoundTag;
