+++ src/pocketmine/block/tile/Sign.php
@@ -21,7 +21,7 @@
 
 declare(strict_types=1);
 
-namespace pocketmine\tile;
+namespace pocketmine\block\tile;
 
 use pocketmine\block\utils\SignText;
 use pocketmine\math\Vector3;
