+++ src/pocketmine/event/inventory/FurnaceSmeltEvent.php
@@ -23,11 +23,11 @@ declare(strict_types=1);
 
 namespace pocketmine\event\inventory;
 
+use pocketmine\block\tile\Furnace;
 use pocketmine\event\block\BlockEvent;
 use pocketmine\event\Cancellable;
 use pocketmine\event\CancellableTrait;
 use pocketmine\item\Item;
-use pocketmine\tile\Furnace;
 
 class FurnaceSmeltEvent extends BlockEvent implements Cancellable{
 	use CancellableTrait;
