+++ src/pocketmine/inventory/FurnaceInventory.php
@@ -23,9 +23,9 @@ declare(strict_types=1);
 
 namespace pocketmine\inventory;
 
+use pocketmine\block\tile\Furnace;
 use pocketmine\item\Item;
 use pocketmine\network\mcpe\protocol\types\WindowTypes;
-use pocketmine\tile\Furnace;
 
 class FurnaceInventory extends ContainerInventory{
 	/** @var Furnace */
