--- src/pocketmine/tile/Bed.php
@@ -1,64 +1,0 @@
-<?php
-
-/*
- *
- *  ____            _        _   __  __ _                  __  __ ____
- * |  _ \ ___   ___| | _____| |_|  \/  (_)_ __   ___      |  \/  |  _ \
- * | |_) / _ \ / __| |/ / _ \ __| |\/| | | '_ \ / _ \_____| |\/| | |_) |
- * |  __/ (_) | (__|   <  __/ |_| |  | | | | | |  __/_____| |  | |  __/
- * |_|   \___/ \___|_|\_\___|\__|_|  |_|_|_| |_|\___|     |_|  |_|_|
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * @author PocketMine Team
- * @link http://www.pocketmine.net/
- *
- *
-*/
-
-declare(strict_types=1);
-
-namespace pocketmine\tile;
-
-use pocketmine\block\utils\DyeColor;
-use pocketmine\math\Vector3;
-use pocketmine\nbt\tag\ByteTag;
-use pocketmine\nbt\tag\CompoundTag;
-use pocketmine\world\World;
-
-class Bed extends Spawnable{
-	public const TAG_COLOR = "color";
-	/** @var DyeColor */
-	private $color;
-
-	public function __construct(World $world, Vector3 $pos){
-		$this->color = DyeColor::RED();
-		parent::__construct($world, $pos);
-	}
-
-	public function getColor() : DyeColor{
-		return $this->color;
-	}
-
-	public function setColor(DyeColor $color) : void{
-		$this->color = $color;
-		$this->onChanged();
-	}
-
-	public function readSaveData(CompoundTag $nbt) : void{
-		if($nbt->hasTag(self::TAG_COLOR, ByteTag::class)){
-			$this->color = DyeColor::fromMagicNumber($nbt->getByte(self::TAG_COLOR));
-		}
-	}
-
-	protected function writeSaveData(CompoundTag $nbt) : void{
-		$nbt->setByte(self::TAG_COLOR, $this->color->getMagicNumber());
-	}
-
-	protected function addAdditionalSpawnData(CompoundTag $nbt) : void{
-		$nbt->setByte(self::TAG_COLOR, $this->color->getMagicNumber());
-	}
-}
